Dedupe `lib` PSX e_mist_door, e_life_up, e_flea_man
