+++ config/splat.us.stno0.yaml
@@ -68,6 +68,7 @@ segments:
       - [0x2478, .data, e_axe_knight]
       - [0x25CC, .data, e_ouija_table]
       - [0x2618, .data, e_flea_man]
+      - [0x2670, data] # unused data?
       - [0x2680, .data, e_skeleton]
       - [0x2770, .data, e_magically_sealed_door]
       - [0x27E0, data]
