+++ src/st/lib/e_flea_man.c
@@ -1,8 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "lib.h"
 
-INCLUDE_ASM("st/lib/nonmatchings/e_flea_man", CheckFieldCollisionY);
-
-INCLUDE_ASM("st/lib/nonmatchings/e_flea_man", UpdateFacingDirection);
-
-INCLUDE_ASM("st/lib/nonmatchings/e_flea_man", EntityFleaMan);
+#include "../e_flea_man.h"
