+++ src/st/lib/e_life_up.c
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "lib.h"
 
-INCLUDE_ASM("st/lib/nonmatchings/e_life_up", EntityLifeUpSpawn);
+#include "../e_life_up.h"
