+++ src/st/lib/e_mist_door.c
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "lib.h"
 
-// Mist door
-INCLUDE_ASM("st/lib/nonmatchings/e_mist_door", EntityMistDoor);
+#include "../e_mist_door.h"
